role_management/requirements.txt
Django>=3.2
djangorestframework>=3.12.0
# Development dependencies
pytest>=6.0
pytest-django>=4.0
black
isort
flake8
# Build tools
wheel
twine
build
